requirements.txt
phidata
python-dotenv
anthropic
sqlalchemy 
psycopg
pgvector
streamlit
pymupdf
pandas
plotly
beautifulsoup4
google-auth 
google-auth-httplib2 
google-api-python-client